Added a Boolean attribute 'pass'
Pass is a Boolean attribute to indicate whether a move is a pass or not. If pass is false, x, y, and c are all relevant information about where a move is placed. If pass is true, c is important to progress the game, but x and y are not important