Modified function parameters to match standards
connect() requires extra to be an integer from 1-3 to set the difficulty.
x,y,c are expected to be the x and y coordinates and the colour of the last move. Nobody answered me on Slack so if this is supposed to be something different, tell Mitch so he can fix this
AI uses integers for move colour: 1 is black, 2 is white. IF YOU WANT TO USE A DIFFERENT SYSTEM FOR THE REST OF THE APPLICATION, THIS MUST BE CONVERTED ON INPUT AND OUTPUT TO THE AI SERVER.